refactor file lib notification
