+++ src/Expensify.js
@@ -35,6 +35,8 @@ import DeeplinkWrapper from './components/DeeplinkWrapper';
 // This lib needs to be imported, but it has nothing to export since all it contains is an Onyx connection
 // eslint-disable-next-line no-unused-vars
 import UnreadIndicatorUpdater from './libs/UnreadIndicatorUpdater';
+// eslint-disable-next-line no-unused-vars
+import subscribePushNotification from './libs/Notification/PushNotification/subscribePushNotification';
 
 Onyx.registerLogger(({level, message}) => {
     if (level === 'alert') {
