+++ src/libs/Notification/PushNotification/index.native.js
@@ -1,17 +1,58 @@
+import _ from 'underscore';
 import Onyx from 'react-native-onyx';
 import Airship, {EventType} from '@ua/react-native-airship';
+import lodashGet from 'lodash/get';
 import Log from '../../Log';
+import NotificationType from './NotificationType';
 import * as PushNotification from '../../actions/PushNotification';
 import ONYXKEYS from '../../../ONYXKEYS';
 import configureForegroundNotifications from './configureForegroundNotifications';
-import subscribeToReportCommentPushNotifications from './subscribeToReportCommentPushNotifications';
-import * as PushNotificationListener from './PushNotificationListener';
 
 let isUserOptedInToPushNotifications = false;
 Onyx.connect({
     key: ONYXKEYS.PUSH_NOTIFICATIONS_ENABLED,
     callback: (val) => (isUserOptedInToPushNotifications = val),
 });
+
+const notificationEventActionMap = {};
+
+/**
+ * Handle a push notification event, and trigger and bound actions.
+ *
+ * @param {String} eventType
+ * @param {Object} notification
+ */
+function pushNotificationEventCallback(eventType, notification) {
+    const actionMap = notificationEventActionMap[eventType] || {};
+    let payload = lodashGet(notification, 'extras.payload');
+
+    // On Android, some notification payloads are sent as a JSON string rather than an object
+    if (_.isString(payload)) {
+        payload = JSON.parse(payload);
+    }
+
+    Log.info(`[PushNotification] Callback triggered for ${eventType}`);
+
+    if (!payload) {
+        Log.warn('[PushNotification] Notification has null or undefined payload, not executing any callback.');
+        return;
+    }
+
+    if (!payload.type) {
+        Log.warn('[PushNotification] No type value provided in payload, not executing any callback.');
+        return;
+    }
+
+    const action = actionMap[payload.type];
+    if (!action) {
+        Log.warn('[PushNotification] No callback set up: ', {
+            event: eventType,
+            notificationType: payload.type,
+        });
+        return;
+    }
+    action(payload);
+}
 
 /**
  * Check if a user is opted-in to push notifications on this device and update the `pushNotificationsEnabled` NVP accordingly.
@@ -43,13 +84,13 @@ function init() {
             PushNotification.setPushNotificationOptInStatus(true);
         }
 
-        PushNotificationListener.pushNotificationEventCallback(EventType.PushReceived, notification.pushPayload);
+        pushNotificationEventCallback(EventType.PushReceived, notification.pushPayload);
     });
 
     // Note: the NotificationResponse event has a nested PushReceived event,
     // so event.notification refers to the same thing as notification above ^
     Airship.addListener(EventType.NotificationResponse, (event) => {
-        PushNotificationListener.pushNotificationEventCallback(EventType.NotificationResponse, event.pushPayload);
+        pushNotificationEventCallback(EventType.NotificationResponse, event.pushPayload);
     });
 
     // Keep track of which users have enabled push notifications via an NVP.
@@ -98,37 +139,58 @@ function deregister() {
 }
 
 /**
+ * Bind a callback to a push notification of a given type.
+ * See https://github.com/Expensify/Web-Expensify/blob/main/lib/MobilePushNotifications.php for the various
+ * types of push notifications sent, along with the data that they provide.
+ *
+ * Note: This implementation allows for only one callback to be bound to an Event/Type pair. For example,
+ *       if we attempt to bind two callbacks to the PushReceived event for reportComment notifications,
+ *       the second will overwrite the first.
+ *
+ * @param {String} notificationType
+ * @param {Function} callback
+ * @param {String} [triggerEvent] - The event that should trigger this callback. Should be one of UrbanAirship.EventType
+ */
+function bind(notificationType, callback, triggerEvent) {
+    if (!notificationEventActionMap[triggerEvent]) {
+        notificationEventActionMap[triggerEvent] = {};
+    }
+    notificationEventActionMap[triggerEvent][notificationType] = callback;
+}
+
+/**
+ * Bind a callback to be executed when a push notification of a given type is received.
+ *
+ * @param {String} notificationType
+ * @param {Function} callback
+ */
+function onReceived(notificationType, callback) {
+    bind(notificationType, callback, EventType.PushReceived);
+}
+
+/**
+ * Bind a callback to be executed when a push notification of a given type is tapped by the user.
+ *
+ * @param {String} notificationType
+ * @param {Function} callback
+ */
+function onSelected(notificationType, callback) {
+    bind(notificationType, callback, EventType.NotificationResponse);
+}
+
+/**
  * Clear all push notifications
  */
 function clearNotifications() {
     Airship.push.clearNotifications();
 }
 
-/**
- * Manage push notification subscriptions on sign-in/sign-out.
- *
- * On Android, AuthScreens unmounts when the app is closed with the back button so we manage the
- * push subscription when the session changes here.
- */
-Onyx.connect({
-    key: ONYXKEYS.NVP_PRIVATE_PUSH_NOTIFICATION_ID,
-    callback: (notificationID) => {
-        if (notificationID) {
-            register(notificationID);
-
-            // Prevent issue where report linking fails after users switch accounts without closing the app
-            init();
-            subscribeToReportCommentPushNotifications();
-        } else {
-            deregister();
-            clearNotifications();
-        }
-    },
-});
-
 export default {
     init,
     register,
     deregister,
+    onReceived,
+    onSelected,
+    TYPE: NotificationType,
     clearNotifications,
 };
